[1.6.1] Split server.rb in several subclasses
