+++ sig/lib/docscribe/server/client.rbs
