+++ sig/lib/docscribe/server/daemon.rbs
