+++ sig/lib/docscribe/server/protocol.rbs
