+++ spec/support/suppress_error_helper.rb
@@ -1,6 +1,11 @@
 # frozen_string_literal: true
 
 module SuppressErrorHelper
+  # Method documentation.
+  #
+  # @raise [StandardError]
+  # @return [Object]
+  # @return [nil] if StandardError
   def suppress_error
     yield
   rescue StandardError
